chore: move EcalcValidationError to common
