+++ src/libecalc/domain/hydrocarbon_export.py
@@ -1,10 +1,10 @@
 from datetime import datetime
 
+from libecalc.common.errors.ecalc_validation_error import EcalcValidationException
 from libecalc.common.time_utils import Period
 from libecalc.common.units import Unit
 from libecalc.common.utils.rates import RateType, TimeSeriesRate
 from libecalc.common.variables import ExpressionEvaluator
-from libecalc.domain.ecalc_validation_error import EcalcValidationException
 from libecalc.domain.regularity import Regularity
 from libecalc.expression.expression import ExpressionType, InvalidExpressionError
 from libecalc.expression.temporal_expression import TemporalExpression
