+++ src/libecalc/domain/infrastructure/energy_components/generator_set/generator_set_validator.py
@@ -1,8 +1,8 @@
-from libecalc.common.errors.exceptions import InvalidColumnException
-from libecalc.domain.ecalc_validation_error import (
+from libecalc.common.errors.ecalc_validation_error import (
     GeneratorSetEqualLengthValidationException,
     GeneratorSetHeaderValidationException,
 )
+from libecalc.common.errors.exceptions import InvalidColumnException
 from libecalc.domain.resource import Resource
