+++ src/libecalc/domain/infrastructure/energy_components/legacy_consumer/system/operational_setting.py
@@ -5,8 +5,8 @@ from typing import Any
 import numpy as np
 from numpy.typing import NDArray
 
+from libecalc.common.errors.ecalc_validation_error import EcalcValidationException
 from libecalc.common.logger import logger
-from libecalc.domain.ecalc_validation_error import EcalcValidationException
 from libecalc.domain.time_series_flow_rate import TimeSeriesFlowRate
 from libecalc.domain.time_series_fluid_density import TimeSeriesFluidDensity
 from libecalc.domain.time_series_pressure import TimeSeriesPressure
