+++ src/libecalc/domain/process/compressor/core/sampled/compressor_model_sampled.py
@@ -7,15 +7,15 @@ from scipy.interpolate import interp1d
 
 from libecalc.common.consumption_type import ConsumptionType
 from libecalc.common.energy_usage_type import EnergyUsageType
-from libecalc.common.errors.exceptions import InvalidColumnException
-from libecalc.common.list.list_utils import array_to_list
-from libecalc.common.logger import logger
-from libecalc.common.units import Unit
-from libecalc.domain.ecalc_validation_error import (
+from libecalc.common.errors.ecalc_validation_error import (
     ProcessEqualLengthValidationException,
     ProcessMissingVariableValidationException,
     ProcessNegativeValuesValidationException,
 )
+from libecalc.common.errors.exceptions import InvalidColumnException
+from libecalc.common.list.list_utils import array_to_list
+from libecalc.common.logger import logger
+from libecalc.common.units import Unit
 from libecalc.domain.process.compressor.core.sampled.compressor_model_sampled_1d import CompressorModelSampled1D
 from libecalc.domain.process.compressor.core.sampled.compressor_model_sampled_2d import (
     CompressorModelSampled2DPsPd,
