+++ src/libecalc/domain/process/compressor/core/train/compressor_train_common_shaft.py
@@ -3,16 +3,16 @@ from copy import deepcopy
 from functools import partial
 from typing import Self
 
+from libecalc.common.errors.ecalc_validation_error import (
+    ProcessChartTypeValidationException,
+    ProcessDischargePressureValidationException,
+)
 from libecalc.common.errors.exceptions import IllegalStateException
 from libecalc.common.fixed_speed_pressure_control import FixedSpeedPressureControl
 from libecalc.common.logger import logger
 from libecalc.common.numeric_methods import (
     find_root,
     maximize_x_given_boolean_condition_function,
-)
-from libecalc.domain.ecalc_validation_error import (
-    ProcessChartTypeValidationException,
-    ProcessDischargePressureValidationException,
 )
 from libecalc.domain.process.compressor.core.results import (
     CompressorTrainResultSingleTimeStep,
