+++ src/libecalc/domain/regularity.py
@@ -1,11 +1,11 @@
 from datetime import datetime
 from typing import Self
 
+from libecalc.common.errors.ecalc_validation_error import EcalcValidationException
 from libecalc.common.time_utils import Period
 from libecalc.common.units import Unit
 from libecalc.common.utils.rates import TimeSeriesFloat
 from libecalc.common.variables import ExpressionEvaluator
-from libecalc.domain.ecalc_validation_error import EcalcValidationException
 from libecalc.expression.expression import ExpressionType, InvalidExpressionError
 from libecalc.expression.temporal_expression import TemporalExpression
