+++ src/libecalc/presentation/yaml/mappers/fluid_mapper.py
@@ -1,4 +1,4 @@
-from libecalc.domain.ecalc_validation_error import EcalcValidationException
+from libecalc.common.errors.ecalc_validation_error import EcalcValidationException
 from libecalc.presentation.yaml.yaml_keywords import EcalcYamlKeywords
 from libecalc.presentation.yaml.yaml_types.models.yaml_fluid import YamlCompositionFluidModel, YamlPredefinedFluidModel
 from libecalc.process.fluid_stream.fluid_model import EoSModel, FluidComposition, FluidModel
