+++ src/libecalc/presentation/yaml/mappers/process_simulation_mapper.py
@@ -1,11 +1,11 @@
 from collections.abc import Sequence
 from typing import Literal, assert_never
 
+from libecalc.common.errors.ecalc_validation_error import EcalcValidationException
 from libecalc.common.errors.exceptions import InvalidResourceException
 from libecalc.common.time_utils import Period
 from libecalc.common.units import Unit
 from libecalc.common.variables import ExpressionEvaluator
-from libecalc.domain.ecalc_validation_error import EcalcValidationException
 from libecalc.domain.process.value_objects.chart.chart import ChartData
 from libecalc.domain.regularity import Regularity
 from libecalc.domain.resource import Resources
