+++ tests/libecalc/domain/process/compressor/test_compressor_train.py
@@ -1,8 +1,8 @@
 import pytest
 from inline_snapshot import snapshot
 
+from libecalc.common.errors.ecalc_validation_error import ProcessChartTypeValidationException
 from libecalc.common.fixed_speed_pressure_control import FixedSpeedPressureControl
-from libecalc.domain.ecalc_validation_error import ProcessChartTypeValidationException
 from libecalc.domain.process.compressor.core.train.compressor_train_common_shaft import CompressorTrainCommonShaft
 from libecalc.process.shaft import SingleSpeedShaft
