+++ tests/libecalc/dto/test_generator_set.py
@@ -6,11 +6,11 @@ import pytest
 from inline_snapshot import snapshot
 
 from libecalc.common.component_type import ComponentType
-from libecalc.common.temporal_model import TemporalModel
-from libecalc.common.time_utils import Period
-from libecalc.domain.ecalc_validation_error import (
+from libecalc.common.errors.ecalc_validation_error import (
     GeneratorSetHeaderValidationException,
 )
+from libecalc.common.temporal_model import TemporalModel
+from libecalc.common.time_utils import Period
 from libecalc.domain.infrastructure.energy_components.generator_set import GeneratorSetModel
 from libecalc.domain.infrastructure.energy_components.generator_set.generator_set_component import (
     GeneratorSetEnergyComponent,
