+++ tests/libecalc/presentation/yaml/domain/test_time_series_pressure.py
@@ -2,8 +2,8 @@ from datetime import datetime
 
 import pytest
 
+from libecalc.common.errors.ecalc_validation_error import ProcessPressureRatioValidationException
 from libecalc.common.time_utils import Period
-from libecalc.domain.ecalc_validation_error import ProcessPressureRatioValidationException
 from libecalc.presentation.yaml.domain.expression_time_series_pressure import (
     ExpressionTimeSeriesPressure,
     InvalidPressureException,
